Fix 1

+++ main.sh
@@ -126,15 +126,18 @@ touch temp_snp_count.txt
 # Check if $snps is a file and filter SNPs accordingly
 if [ -f "$snps" ]; then
     log "	Filtering SNPs based on $snps"
-    awk -v sidx="$snp_id_idx" 'NR==FNR {snps[$1]; next} $sidx in snps' "$snps" "$infile" | sort | uniq -c > temp_snp_count.txt \
-    	|| log "main.sh Error: unable to create filtered snp count file" && exit 1
+    if ! awk -v sidx="$snp_id_idx" 'NR==FNR {snps[$1]; next} $sidx in snps' "$snps" "$infile" | sort | uniq -c > temp_snp_count.txt; then
+        log "main.sh Error: unable to create filtered snp count file" && exit 1
+    fi
 else
-    awk -v sidx="$snp_id_idx" '{ print $sidx }' "$infile" | sort | uniq -c > temp_snp_count.txt \
-        || log "main.sh Error: unable to create snp count file" && exit 1
+    if ! awk -v sidx="$snp_id_idx" '{ print $sidx }' "$infile" | sort | uniq -c > temp_snp_count.txt; then
+        log "main.sh Error: unable to create snp count file" && exit 1
+    fi
 fi
 
-awk '{ print $2, $1 }' temp_snp_count.txt > snp_count.txt \
-    || log "main.sh Error: unable to create snp_count.txt file" && exit 1
+if ! awk '{ print $2, $1 }' temp_snp_count.txt > snp_count.txt; then
+    log "main.sh Error: unable to create snp_count.txt file" && exit 1
+fi
 
 rm temp_snp_count.txt
 log "	Created file snp_count.txt"
